+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow for raw config generation from CLI.
+
 ## [0.2.11] - 2021-04-16
 
 ### Fixed